Commit schema version 5.2.0
Changes:

* "users" is now "account" (singular)
* Removed folder_name constraint: duplicate folders for everyone
* Crucial: removed profile_id from almost everywhere (see below)
* Crucial: removed user_id from location (it's now profile-locked)
* Renamed 'journal.seconds' to 'journal.duration'

I was increasingly uncomfortable with my schema diagram in dbdiagram.io
because entities like "project" could be part of both a profile and a
folder. What's the issue? Well, a folder is *also* part of a profile.
Which meant that, technically, a single project could belong to two
different profiles and that my data integrity would have to be maintained
at the software level. Decided to try this new approach:

* project, activity, location, tag: belong to folders
* folders are always required (there's a hidden "root" folder for all)
* folders belong to a profile
* a profile belongs to a user

This greatly simplified the diagram, removed the "double ownership" from
these key entities. Now, it just needs to be tested.