Add support for account fetch, update, delete
And this is in AccountController, AccountService, and DatabaseService
all in one commit.

This is still subject to further refinements, of course; but I'm just
plowing through my prototype stage here, given how long I've already
spent on this mess. I can't even tell you, because it's spread across
two tkdb files and now an ATracker instance.