Commit massive DRY refactoring [m]
I know I've DRY'd this up already a few times, but THIS IS IT. At this
point, BaseConroller and BaseService have all the common code and the
individual controllers and services are now incredibly thin. For the
controllers, they're now just one-liners. For the services, they're
still mostly the same, except for create() which is now "universal". It
takes a $data array and creates the $payload by extracting keys. It's
less "strict" this way, but way more flexible (which is always the
trade-off in scripting languages, right?)

Additionally, create now returns an array, though in most cases, it has
just one element: 'id'

Oh, and to that end, in the controllers, all 'id' values are now just
$id (or {id}), again, for readability and reusuability.

I also cleaned up some bugs, expanded some endpoints, and introduced
the next entity: Projects. The .http files are coming next.