Stub in initial Nelmio and OpenApi documentation for account creation
I'm going to leave it here for a while too. The API documentation is
largely a distraction at this point: one I don't need. I mainly wanted
to get the infrastructure set up for it, because it might be useful
later (and the completionist in me will want to see it through at some
point).