Commit another refactoring of entity create()
In spite of finalizing this about a week ago in commit 91737f7, I've
now made one more pass at how entity creation works. This was driven by
the introduction of the Nelmio API documentation which itself got me to
introduce a DTO to define the entity creation request payloads.

Quick overview:

* Each entity (well, four of them as of this commit) has a DTO for creation
* The DTO is used by #[MapRequestPayload] to provide true objects instead of arrays
* The BaseController _create() method has been removed: there's no commonality any more
* It now passes the specific object to <Entity>Service's create method
* And create is now back to returning an int instead of an array

Along for the ride: a new ApiResponse::success method to return 201s instead of 200s

This should be the final approach.