Implement database-driven security system
Before: users were defined in-memory and there were only two api_client
and api_admin. And these had the client and admin roles hardcoded to
them in security.yaml. Also: API keys were defined in the .env file
and "softcoded" in a way that would be difficult to manage going forward
(in short: zero scalability, which was fine for API prototyping).

Now: users are defined in the database. And the account table (will soon
have) an API token column (along with an is_admin flag). The auth class
ditches the .env keys and does a database lookup. Then hands all of that
over to Symfony's security system. There's new ApiKeyUserProvider and
ApiKeyUser classes.

There's still some prototype-level hardcoding going on, but the overall
framework is in place. User login will be committed next and then we're
going to be a LOT closer to making this real.