Commit new LoginController and LoginService
A username and password are submitted to the /login endpoint (which is
outside of /api, since that's behind the firewall). We look up the user
and verify the password. If okay, we generate a new token and return it
to the caller. If not, processing ends.

The token is a UUID, so the ramsey package comes along for the ride.