Commit a few initial attempts at permission
There's a new PermissionService, which is just a POC at best. It's very,
very likely this will change or take a completely different direction.

ProfileController is attempting to use the new PermissionService, but
I'm not sold on this just yet.

I've started using the IsGranted annotation.

And the login.http test script is coming along for the ride.