Commit Journal entry import WIP
* The endpoint now contains {profileID} so we know where to import
* Removed $userId and converted $result to $rowCount
* Adding an early test copy of journal-import.http

It's slow, of course, processing a row at a time. Takes a few seconds
to go through ~230 rows in the CSV. But it's working for this still-
early prototyping stage.

Two important notes:

* This is 100% geared towards ATracker import
* Further, *my* version of ATracker (I'm using its tags for project/activity/location)
* It hardcodes id=1 for project, activity, location

I need to create a generic import (and then write an adapter to convert
the custom ATracker CSV into an official TK5 import format. But that will
come later. First up, need to fix the id=1 placeholder.