Commit a solid known-good point for JournalImport
As of this commit, Journal Import can:

* Read a "standard" TK5 CSV formatted file
* Grab start/stop time from it
* Calculate duration on the fly
* Pull in memo
* Accept a "tree" of projects (e.g., "A::B::C::D")
* Fetch or create folders A, B, and C.
* Fetch or create project D
* Do the same for activities, locations, and arrays of tags

And from a data-integrity point of view, the ownership hierachy is
fully enforced:

* Account > Profile > Folder > "Dimensional Entity"
* Every folder must have a parent
* The root-most system folder parents to itself (so id=parent_id will
  have to be used in the future for recursive checks)

Lots and lots of testing to go. Plus, I'll now have to create the
intermediary ATracker CSV -> TK5 CSV converter. But when I get there,
I'll be able to introduce my new (temporary) workflow.